fix display error
